# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/bookings": {
      "post": {
        "operationId": "BookingsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingMutationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "Create a booking owned by the current user",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/my": {
      "get": {
        "operationId": "BookingsController_listMine",
        "parameters": [
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "scheduled",
                "completed",
                "cancelled",
                "all"
              ],
              "type": "string"
            }
          },
          {
            "name": "view",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "upcoming",
                "current",
                "past",
                "cancelled",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingSummaryDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "List current user bookings with view and date/status filters",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/upcoming": {
      "get": {
        "operationId": "BookingsController_upcoming",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingSummaryDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "Read up to five current user active upcoming bookings",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/admin/all": {
      "get": {
        "operationId": "BookingsController_listAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "scheduled",
                "completed",
                "cancelled",
                "all"
              ],
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ownerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingSummaryDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "List every booking for administrators",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/admin/owners": {
      "get": {
        "operationId": "BookingsController_owners",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingOwnerDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "List booking owners for the administrative filter",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/{id}": {
      "get": {
        "operationId": "BookingsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "Read a booking; administrators may read any booking",
        "tags": [
          "bookings"
        ]
      },
      "patch": {
        "operationId": "BookingsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "Change a current user future booking using its optimistic version",
        "tags": [
          "bookings"
        ]
      },
      "delete": {
        "operationId": "BookingsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCancellationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "Cancel a current user future booking using its optimistic version",
        "tags": [
          "bookings"
        ]
      }
    }
  },
  "info": {
    "title": "Roomly API",
    "description": "HTTP API for the Roomly meeting room booking service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Анна Соколова"
          },
          "email": {
            "type": "string",
            "example": "employee@northstar.local"
          },
          "role": {
            "type": "string",
            "enum": [
              "employee",
              "admin"
            ]
          }
        },
        "required": [
          "name",
          "email",
          "role"
        ]
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "example": 400
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Не удалось обработать запрос"
          },
          "fieldErrors": {
            "type": "string",
            "example": [
              "subject: Укажите тему"
            ]
          },
          "requestId": {
            "type": "string",
            "example": "f026e5ca-01e4-4c79-a52e-c44d9c32d967"
          }
        },
        "required": [
          "status",
          "code",
          "message",
          "fieldErrors",
          "requestId"
        ]
      },
...
      "BookingMutationDto": {
        "type": "object",
        "properties": {
          "roomId": {
            "type": "string",
            "example": "room-atlas"
          },
          "subject": {
            "type": "string",
            "example": "Планирование"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "participants": {
            "type": "number",
            "example": 3,
            "minimum": 1
          },
          "date": {
            "type": "string",
            "example": "2026-10-05"
          },
          "start": {
            "type": "string",
            "example": "10:00"
          },
          "end": {
            "type": "string",
            "example": "11:00"
          }
        },
        "required": [
          "roomId",
          "subject",
          "participants",
          "date",
          "start",
          "end"
        ]
      },
      "BookingRoomDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "room-atlas"
          },
          "name": {
            "type": "string",
            "example": "Атлас"
          },
          "floor": {
            "type": "number"
          },
          "location": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "unavailable"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "floor",
          "location",
          "capacity",
          "status"
        ]
      },
      "CancellationDto": {
        "type": "object",
        "properties": {
          "cancelledAt": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "type": "string",
            "enum": [
              "owner",
              "admin"
            ]
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "actor": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "required": [
          "cancelledAt",
          "type",
          "actor"
        ]
      },
      "BookingDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "booking-123"
          },
          "subject": {
            "type": "string",
            "example": "Планирование"
          },
          "room": {
            "$ref": "#/components/schemas/BookingRoomDto"
          },
          "date": {
            "type": "string",
            "example": "2026-10-05"
          },
          "start": {
            "type": "string",
            "example": "10:00"
          },
          "end": {
            "type": "string",
            "example": "11:00"
          },
          "participants": {
            "type": "number",
            "minimum": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "completed",
              "cancelled"
            ]
          },
          "isCurrent": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "number",
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "owner": {
            "$ref": "#/components/schemas/UserDto"
          },
          "cancellation": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CancellationDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "subject",
          "room",
          "date",
          "start",
          "end",
          "participants",
          "status",
          "isCurrent",
          "version",
          "createdAt",
          "updatedAt",
          "owner"
        ]
      },
      "BookingSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "booking-123"
          },
          "subject": {
            "type": "string",
            "example": "Планирование"
          },
          "room": {
            "$ref": "#/components/schemas/BookingRoomDto"
          },
          "date": {
            "type": "string",
            "example": "2026-10-05"
          },
          "start": {
            "type": "string",
            "example": "10:00"
          },
          "end": {
            "type": "string",
            "example": "11:00"
          },
          "participants": {
            "type": "number",
            "minimum": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "completed",
              "cancelled"
            ]
          },
          "isCurrent": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "subject",
          "room",
          "date",
          "start",
          "end",
          "participants",
          "status",
          "isCurrent"
        ]
      },
      "BookingOwnerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Анна Соколова"
          },
          "email": {
            "type": "string",
            "example": "employee@northstar.local"
          },
          "role": {
            "type": "string",
            "enum": [
              "employee",
              "admin"
            ]
          },
          "id": {
            "type": "string",
            "example": "user-employee"
          }
        },
        "required": [
          "name",
          "email",
          "role",
          "id"
        ]
      },
      "BookingUpdateDto": {
        "type": "object",
        "properties": {
          "roomId": {
            "type": "string",
            "example": "room-atlas"
          },
          "subject": {
            "type": "string",
            "example": "Планирование"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "participants": {
            "type": "number",
            "example": 3,
            "minimum": 1
          },
          "date": {
            "type": "string",
            "example": "2026-10-05"
          },
          "start": {
            "type": "string",
            "example": "10:00"
          },
          "end": {
            "type": "string",
            "example": "11:00"
          },
          "version": {
            "type": "number",
            "example": 1,
            "minimum": 1
          }
        },
        "required": [
          "roomId",
          "subject",
          "participants",
          "date",
          "start",
          "end",
          "version"
        ]
      },
      "BookingCancellationDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "reason": {
            "type": "string",
            "example": "Плановый ремонт",
            "maxLength": 1000
          }
        },
        "required": [
          "version"
        ]
      }
    }
  }
}